IITP_Naxus/requirements.txt
Flask==2.1.1
Flask-Cors==5.0.1
Flask-Mail==0.9.1
requests==2.26.0
SQLAlchemy==1.4.31
gunicorn==20.1.0
python-dotenv==0.19.2
